+++ Bill_Software/corporate/business/print/NewChhalanDuplicate.aspx.cs
@@ -19,7 +19,7 @@ namespace Bill_Software.corporate.business.print
         DataTable dtRep = new DataTable();
         DataTable dtChadd = new DataTable();
         StringBuilder strProduct = new StringBuilder();
-        public int TQ = 0;
+        public decimal TQ = 0;
         protected void Page_Load(object sender, EventArgs e)
         {
             string Chalan_No = Request.QueryString["Chalan_No"];
@@ -170,7 +170,7 @@ namespace Bill_Software.corporate.business.print
                     string Product_id = dtChPro.Rows[i]["Product_id"].ToString();
                     string HSN = dtChPro.Rows[i]["Product_code"].ToString();
                     string Product_name = dtChPro.Rows[i]["Product_name"].ToString();
-                    int Quantity = Convert.ToInt32(dtChPro.Rows[i]["Quantity"]);
+                    decimal Quantity = Convert.ToInt32(dtChPro.Rows[i]["Quantity"]);
                     string Unit = dtChPro.Rows[i]["Unit"].ToString();
                     string Specification = dtChPro.Rows[i]["specification"].ToString();
                     string ItemNo = dtChPro.Rows[i]["ItemNo"].ToString();
